requirements.txt
Django==2.2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeltranslation==0.16.1
djangorestframework==3.12.1
Faker==5.0.2
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
